Initial commit with the project implemented
Sorry I implemented it without git in a single sitting so that it doesn't occupy much of my time and didn't write git logs of the process because the code is terrible eitherway. I do not recommend trying to learn anything with this repository. It's just a web tool quickly hacked together.